Fixing warnings when using clang to compile.

+++ Peach.Core.Analysis.Pin.BasicBlocks/bblocks.cpp
@@ -75,12 +75,12 @@ VOID PIN_FAST_ANALYSIS_CALL rememberBlock(ADDRINT bbl)
 	ret = setKnownBlocks.insert(bbl);
 	if(ret.second == true)
 	{
-		fprintf(trace, "%16.16zx\n", bbl);
+		fprintf(trace, "%16.16zx\n", (size_t)bbl);
 		fflush(trace);
 
 		if(haveExisting)
 		{
-			fprintf(existing, "%16.16zx\n", bbl);
+			fprintf(existing, "%16.16zx\n", (size_t)bbl);
 			fflush(existing);
 		}
 	}
